requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
lipsum==0.1.2
MarkupSafe==3.0.2
matplotlib==3.10.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
pillow==11.1.0
pluggy==1.5.0
pycryptodome==3.21.0
pyparsing==3.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
ruff==0.9.7
setuptools==75.8.0
simplejson==3.20.1
six==1.17.0
types-pytz==2025.1.0.20250204
types-PyYAML==6.0.12.20241230
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
